requirements.txt
python-telegram-bot==20.7
Flask==3.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
googlemaps==4.10.0
yandex-maps==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
PyJWT==2.8.0
aiohttp==3.9.1
alembic==1.13.1
bcrypt==4.1.2
